some cleanup and renaming ... still not 100% happy
